update youtubedl-gui to use yt-dlp instead of youtube-dl
Since this package is intended for sid (and bookworm), we're
not going to complicate things by checking for yt-dlp or youtube-dl
at runtime. Instead, just call out to yt-dlp with the same options
that were used with youtube-dl.

It's also worth noting that at least in my tests, yt-dlp downloads
videos *significantly* faster than youtube-dl; another reason not to
bother checking for youtube-dl.